STRINGS! and std?...
Ahem... okay so the promised strings aren't in the way I promised they would be... theese strings are c strings... perfect for the C ABI pain to work with as a user. But i'll prob make em work next commit in my own way, and type system.
Also changed the way to compile it to an object, the fact that gcc / clang can still be used to compile the object files into binary if you wish